+++ adminforth/spa/src/App.vue
@@ -173,7 +173,7 @@
     </aside>
     
 
-    <div class="sm:ml-64" v-if="loggedIn && routerIsReady && loginRedirectCheckIsReady">
+    <div class="sm:ml-64 max-w-[100vw]" v-if="loggedIn && routerIsReady && loginRedirectCheckIsReady">
       <div class="p-0 dark:border-gray-700 mt-14">
         <RouterView/>     
       </div>